+++ Cargo.toml
@@ -35,6 +35,7 @@ members = [
     "crates/versions",
     "hotshot-events-service",
     "hotshot-query-service",
+    "hotshot-query-service-types",
     "hotshot-state-prover",
     "light-client",
     "light-client-query-service",
@@ -103,7 +104,15 @@ authors = ["Espresso Systems <hello@espressosys.com>"]
 edition = "2024"
 
 [workspace.dependencies]
-alloy = { version = "1.0", features = ["eips", "provider-ws", "json-rpc", "signer-mnemonic", "node-bindings", "getrandom", "signer-ledger"] }
+alloy = { version = "1.0", features = [
+    "eips",
+    "provider-ws",
+    "json-rpc",
+    "signer-mnemonic",
+    "node-bindings",
+    "getrandom",
+    "signer-ledger",
+] }
 alloy-compat = { path = "alloy-compat" }
 anyhow = "1.0"
 ark-bls12-381 = "0.5"
@@ -139,7 +148,9 @@ byteorder = { version = "1" }
 bytes = { version = "1.11.0", features = ["serde"] }
 bytesize = "1.3"
 cbor4ii = { version = "1.0", features = ["serde1"] }
-cdn-broker = { git = "https://github.com/EspressoSystems/Push-CDN", tag = "0.6.2-bcompat-patch2", package = "cdn-broker", features = ["global-permits"] }
+cdn-broker = { git = "https://github.com/EspressoSystems/Push-CDN", tag = "0.6.2-bcompat-patch2", package = "cdn-broker", features = [
+    "global-permits",
+] }
 cdn-client = { git = "https://github.com/EspressoSystems/Push-CDN", tag = "0.6.2-bcompat-patch2", package = "cdn-client" }
 cdn-marshal = { git = "https://github.com/EspressoSystems/Push-CDN", tag = "0.6.2-bcompat-patch2", package = "cdn-marshal" }
 chrono = { version = "0.4", features = ["serde"] }
@@ -151,7 +162,9 @@ cld = "0.5"
 client = { path = "client" }
 cliquenet = { path = "crates/cliquenet", features = ["metrics"] }
 coarsetime = "0.1.34"
-committable = { git = "https://github.com/EspressoSystems/commit.git", features = ["ark-serialize"] }
+committable = { git = "https://github.com/EspressoSystems/commit.git", features = [
+    "ark-serialize",
+] }
 criterion = "0.8.1"
 csv = "1"
 custom_debug = "0.6"
@@ -198,6 +211,7 @@ hotshot-macros = { path = "crates/hotshot/macros" }
 hotshot-new-protocol = { path = "crates/hotshot/new-protocol" }
 hotshot-orchestrator = { path = "crates/hotshot/orchestrator" }
 hotshot-query-service = { path = "hotshot-query-service" }
+hotshot-query-service-types = { path = "hotshot-query-service-types" }
 hotshot-state-prover = { path = "hotshot-state-prover" }
 hotshot-task = { path = "crates/hotshot/task" }
 hotshot-task-impls = { path = "crates/hotshot/task-impls" }
@@ -209,23 +223,69 @@ include_dir = "0.7"
 indexmap = { version = "2", features = ["serde"] }
 insta = { version = "1.43", features = ["yaml"] }
 itertools = "0.14"
-jf-advz = { git = "https://github.com/EspressoSystems/jellyfish-compat", tag = "jf-advz-v0.2.4", features = ["std", "parallel"] }
+jf-advz = { git = "https://github.com/EspressoSystems/jellyfish-compat", tag = "jf-advz-v0.2.4", features = [
+    "std",
+    "parallel",
+] }
 jf-crhf = { git = "https://github.com/EspressoSystems/jellyfish", tag = "jf-crhf-v0.2.0" }
 jf-merkle-tree = { git = "https://github.com/EspressoSystems/jellyfish", tag = "jf-merkle-tree-v0.4.0" }
-jf-merkle-tree-compat = { git = "https://github.com/EspressoSystems/jellyfish-compat", tag = "jf-merkle-tree-compat-v0.3.1", features = ["std"] }
-jf-pcs = { git = "https://github.com/EspressoSystems/jellyfish", tag = "jf-pcs-v0.3.0", features = ["std", "parallel"] }
-jf-plonk = { git = "https://github.com/EspressoSystems/jellyfish", tag = "jf-plonk-v0.7.0", features = ["test-apis"] }
-jf-plonk-compat = { package = "jf-plonk", git = "https://github.com/EspressoSystems/jellyfish", branch = "jf-relation-v0.5.0-compat", features = ["std"] }
-jf-poseidon2 = { git = "https://github.com/EspressoSystems/jellyfish", tag = "jf-poseidon2-v0.2.0", features = ["bn254"] }
-jf-relation = { git = "https://github.com/EspressoSystems/jellyfish", tag = "jf-relation-v0.5.0", features = ["std"] }
-jf-relation-compat = { package = "jf-relation", git = "https://github.com/EspressoSystems/jellyfish", branch = "jf-relation-v0.5.0-compat", features = ["std"] }
-jf-rescue = { git = "https://github.com/EspressoSystems/jellyfish", tag = "jf-rescue-v0.3.0", features = ["std", "parallel", "gadgets"] }
-jf-rescue-compat = { package = "jf-rescue", git = "https://github.com/EspressoSystems/jellyfish", branch = "jf-relation-v0.5.0-compat", features = ["gadgets"] }
-jf-signature = { git = "https://github.com/EspressoSystems/jellyfish", tag = "jf-signature-v0.4.0", features = ["std", "bls", "schnorr", "gadgets"] }
-jf-signature-compat = { package = "jf-signature", git = "https://github.com/EspressoSystems/jellyfish", branch = "jf-relation-v0.5.0-compat", features = ["schnorr", "bls", "gadgets"] }
+jf-merkle-tree-compat = { git = "https://github.com/EspressoSystems/jellyfish-compat", tag = "jf-merkle-tree-compat-v0.3.1", features = [
+    "std",
+] }
+jf-pcs = { git = "https://github.com/EspressoSystems/jellyfish", tag = "jf-pcs-v0.3.0", features = [
+    "std",
+    "parallel",
+] }
+jf-plonk = { git = "https://github.com/EspressoSystems/jellyfish", tag = "jf-plonk-v0.7.0", features = [
+    "test-apis",
+] }
+jf-plonk-compat = { package = "jf-plonk", git = "https://github.com/EspressoSystems/jellyfish", branch = "jf-relation-v0.5.0-compat", features = [
+    "std",
+] }
+jf-poseidon2 = { git = "https://github.com/EspressoSystems/jellyfish", tag = "jf-poseidon2-v0.2.0", features = [
+    "bn254",
+] }
+jf-relation = { git = "https://github.com/EspressoSystems/jellyfish", tag = "jf-relation-v0.5.0", features = [
+    "std",
+] }
+jf-relation-compat = { package = "jf-relation", git = "https://github.com/EspressoSystems/jellyfish", branch = "jf-relation-v0.5.0-compat", features = [
+    "std",
+] }
+jf-rescue = { git = "https://github.com/EspressoSystems/jellyfish", tag = "jf-rescue-v0.3.0", features = [
+    "std",
+    "parallel",
+    "gadgets",
+] }
+jf-rescue-compat = { package = "jf-rescue", git = "https://github.com/EspressoSystems/jellyfish", branch = "jf-relation-v0.5.0-compat", features = [
+    "gadgets",
+] }
+jf-signature = { git = "https://github.com/EspressoSystems/jellyfish", tag = "jf-signature-v0.4.0", features = [
+    "std",
+    "bls",
+    "schnorr",
+    "gadgets",
+] }
+jf-signature-compat = { package = "jf-signature", git = "https://github.com/EspressoSystems/jellyfish", branch = "jf-relation-v0.5.0-compat", features = [
+    "schnorr",
+    "bls",
+    "gadgets",
+] }
 jf-utils = { git = "https://github.com/EspressoSystems/jellyfish", tag = "jf-utils-v0.5.0" }
 lazy_static = "1"
-libp2p = { package = "libp2p", version = "0.56", default-features = false, features = ["macros", "autonat", "cbor", "dns", "gossipsub", "identify", "kad", "quic", "request-response", "secp256k1", "serde", "tokio"] }
+libp2p = { package = "libp2p", version = "0.56", default-features = false, features = [
+    "macros",
+    "autonat",
+    "cbor",
+    "dns",
+    "gossipsub",
+    "identify",
+    "kad",
+    "quic",
+    "request-response",
+    "secp256k1",
+    "serde",
+    "tokio",
+] }
 libp2p-identity = { version = "0.2", features = ["ed25519", "serde"] }
 libp2p-swarm-derive = { version = "0.35" }
 light-client = { path = "light-client", features = ["clap"] }
@@ -281,7 +341,12 @@ sha2 = "0.10"
 sha3 = "0.10.8"
 snafu = "0.8"
 snow = { version = "0.10.0", features = ["ring-accelerated"] }
-sqlx = { version = "=0.8.3", features = ["bit-vec", "postgres", "runtime-tokio", "tls-native-tls"] }
+sqlx = { version = "=0.8.3", features = [
+    "bit-vec",
+    "postgres",
+    "runtime-tokio",
+    "tls-native-tls",
+] }
 staking-cli = { path = "staking-cli" }
 static_assertions = "1.1"
 strum = { version = "0.26", features = ["derive"] }
