+++ hotshot-query-service-types/src/explorer/errors.rs
