+++ hotshot-query-service-types/src/explorer/traits.rs
