+++ light-client/Cargo.toml
@@ -7,7 +7,12 @@ edition.workspace = true
 [features]
 default = ["decaf"]
 decaf = []
-testing = ["espresso-types/testing", "bitvec", "hotshot-contract-adapter", "rand"]
+testing = [
+    "espresso-types/testing",
+    "bitvec",
+    "hotshot-contract-adapter",
+    "rand",
+]
 
 [dependencies]
 alloy = { workspace = true }
@@ -19,7 +24,7 @@ derivative = { workspace = true }
 derive_more = { workspace = true }
 espresso-types = { workspace = true }
 futures = { workspace = true }
-hotshot-query-service = { workspace = true }
+hotshot-query-service-types = { workspace = true }
 hotshot-types = { workspace = true }
 jf-advz = { workspace = true }
 jf-merkle-tree-compat = { workspace = true }
@@ -39,7 +44,9 @@ hotshot-contract-adapter = { workspace = true, optional = true }
 rand = { workspace = true, optional = true }
 
 [dev-dependencies]
-espresso-node = { workspace = true, default-features = false, features = ["testing"] }
+espresso-node = { workspace = true, default-features = false, features = [
+    "testing",
+] }
 light-client = { path = ".", features = ["testing"] }
 pretty_assertions = { workspace = true }
 test-log = { workspace = true }
