+++ light-client/src/testing.rs
@@ -19,7 +19,7 @@ use espresso_types::{
     v0_3::{AuthenticatedValidator, RegisteredValidator, StakeTableEvent},
 };
 use hotshot_contract_adapter::sol_types::StakeTableV2::{Delegated, ValidatorRegistered};
-use hotshot_query_service::{
+use hotshot_query_service_types::{
     availability::{LeafHash, LeafId, LeafQueryData},
     node::{BlockHash, BlockId},
 };
